requirements.txt
# Core RAG Framework
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma

# Document Parsing
llama-index-readers-docling
docling
pypdf
python-pptx

# Vector Database
chromadb

# Visualization
plotly

# Data Schema & Config
pydantic>=2.0
pyyaml
python-dotenv

# Data Processing
pandas
openpyxl
tabulate

# Notebook
jupyter
ipywidgets
